remove some useless files and update the scripts and README files.

+++ 02-config.sh
@@ -1,12 +1,14 @@
-Red='\e[0;31m';
-Gre='\e[0;32m';
-Whi='\e[0;37m';
+Red='\e[0;31m'
+Gre='\e[0;32m'
+Whi='\e[0;37m'
 
 echo -e "[${Gre}*${Whi}] Setting up config files"
 
 echo -e "➞ [${Red}*${Whi}] Removing ~/dotfiles-backup for backing up current configs"
 
-rm -rf ~/dotfiles-backup
+if [-d ~/dotfiles-backup]; then
+	rm -rf ~/dotfiles-backup
+fi
 
 echo -e "➞ [${Gre}*${Whi}] Backing up existing directories"
 
@@ -18,20 +20,17 @@ mkdir -p ~/dotfiles-backup/home
 
 cd ~/dotfiles
 
-for dir in $(find .config -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}')
-do
+for dir in $(find .config -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}'); do
 	mv ~/.config/$dir ~/dotfiles-backup/config
 done
 
-for dir in $(find .config -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}')
-do
+for dir in $(find .config -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}'); do
 	mkdir -p ~/.config/$dir
 done
 
 # --------
 
-for dir in $(find home -mindepth 1 -maxdepth 1 | awk -F "/" '{print $NF}')
-do
+for dir in $(find home -mindepth 1 -maxdepth 1 | awk -F "/" '{print $NF}'); do
 	mv ~/$dir ~/dotfiles-backup/home
 done
 
@@ -43,8 +42,7 @@ echo -e "➞ [${Gre}*${Whi}] Symlinking directories in .config"
 
 cd ~/dotfiles/.config
 
-for dir in $(find . -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}')
-do
+for dir in $(find . -maxdepth 1 -mindepth 1 | awk -F "/" '{print $NF}'); do
 	stow -t ~/.config/$dir $dir
 done
